containerised tab1 components
