+++ meteor-app/imports/ui/service/service.js
@@ -2,12 +2,14 @@ import React from 'react'
 import SearchBar from './service-item-search'
 import ServiceItemTag from './serviceItem-tag'
 import { ServiceContextProvider } from './service-context'
+import ServiceItemTagContainer from './serviceItem-tag-container'
+import ServiceItemSearchContainer from './service-item-search-container'
 
 function Service() {
   return (
     <ServiceContextProvider>
-      <SearchBar />
-      <ServiceItemTag />
+      <ServiceItemSearchContainer />
+      <ServiceItemTagContainer />
     </ServiceContextProvider>
   )
 }
