+++ meteor-app/imports/ui/service/serviceItem-tag.js
@@ -1,32 +1,8 @@
-import React, { useState, useContext } from 'react'
+import React from 'react'
 import './serviceItem-tag.css'
-import { ServiceContext } from './service-context'
 
 export default function ServiceItemTag(props) {
-  const [state, setState] = useContext(ServiceContext)
-  const tags = state.tags
-
-  function removeTag(index) {
-    const newTags = [...tags]
-    newTags.splice(index, 1)
-    const newState = { ...state, tags: newTags }
-    setState(newState)
-  }
-
-  function toggleTag(item) {
-    const newTags = [...tags]
-    newTags.map(tag => {
-      tag.items
-        ? tag.items.map(serviceItem => {
-            if (serviceItem.title === item.title) {
-              serviceItem.greyed = !serviceItem.greyed
-            }
-          })
-        : null
-    })
-    const newState = { ...state, tags: newTags }
-    setState(newState)
-  }
+  const { removeTag, toggleTag, tags } = props
 
   return (
     <div className="wrapper">
