+++ meteor-app/package.json
@@ -88,12 +88,8 @@
     "react-data-grid": "^6.1.0",
     "react-data-grid-addons": "^6.1.0",
     "react-datepicker": "^2.10.1",
-<<<<<<< HEAD
     "react-dom": "^16.13.0",
-=======
-    "react-dom": "^16.8.6",
     "react-hook-form": "^4.10.1",
->>>>>>> d40450e1b93167355c5e69038509e90fdca488ab
     "react-jsonschema-form-semanticui": "^1.0.3",
     "react-modal": "^3.1.13",
     "react-radio-group": "^3.0.2",
